// src/app/Models/provider.model.ts
export class ProviderModel {
    id: number;
    provider_number: string;
    npi: string;
    first_name: string;
    middle_name: string;
    last_name: string;
    extension_name: string;
    qualifications: string;
    gender: string;
    main_phone: string;
    direct_dial: string;
    fax: string;
    email: string;
    web_url: string;
    facility_id: number;
    facility: any;
    profession: string;
    speciality1: string;
    speciality2: string;
    title1: string;
    title2: string;
    title3: string;
    title4: string;
    title5: string;
    status: string;
    date_last_modified: string;
    modified_by: string;
    image_url: string;
}